relocates logic
