--- packages/plugins/src/registry/modules.yaml
@@ -1,1 +1,0 @@
-plugins:
